requirements.txt
pillow==7.0.0
nuitka==6.1.70